# perf: disable video on cypress tests by default