--- a/cypress.config.ts
+++ b/cypress.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
   e2e: {
     supportFile: false,
     defaultCommandTimeout: 10000,
+    video: false,
     setupNodeEvents(on, config) {
       // implement node event listeners here
     },